requirements.txt
langchain
langchain-astradb
langchain-huggingface
langchain-groq
langchain-community
datasets
pypdf
python-dotenv
pandas
flask
prometheus_client
pytest
chardet
scikit-learn
astrapy
transformers
torch
rank-bm25
flask
bs4
requests
selenium
psycopg2-binary
flask-limiter
gunicorn
boto3
redis
flask-cors
